Create Session 1 code
So, this was what I did on the first session.  I know it doesn't look like much, but this was because I forgot a lot of basic commands, ran into a lot of syntax issues, and had to learn the "input('> ')" function.  It was a tough day, but I got off to a good start.